Move validation.ts to lib/guide-validation.ts and fix imports
